TM-1086 : Delete staff role deletes system roles as well - Resolved
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@2813 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ grails-app/controllers/PersonController.groovy
@@ -436,8 +436,12 @@ class PersonController {
 					}
 				}
 				if(params.manageRoles != '0' && params.role){
-					PartyRole.executeUpdate("delete from PartyRole where party = '$personInstance.id'  ")
 					if(params.role){
+                        def existingRoles = PartyRole.findAll("from PartyRole where party = :person and roleType.description like 'staff%' \
+                                                                and roleType.id not in (:roles) group by roleType",[roles:params.role, person:personInstance])?.roleType
+                        if(existingRoles){
+                            PartyRole.executeUpdate("delete from PartyRole where party = '$personInstance.id' and roleType in (:roles)",[roles:existingRoles])
+                        }
 						Set newRoles = []
 						newRoles.addAll(params.role)
 						userPreferenceService.setUserRoles(newRoles, personInstance.id)
